fix: enforce strict resync input framing
